+++ packages/vtable-export/demo/menu.ts
@@ -29,6 +29,14 @@ export const menus = [
       {
         path: 'list',
         name: 'list-large'
+      },
+      {
+        path: 'list',
+        name: 'list-tree'
+      },
+      {
+        path: 'list',
+        name: 'list-checkbox'
       }
     ]
   },