+++ packages/vtable/src/components/tooltip/TooltipHandler.ts
@@ -195,7 +195,8 @@ export class TooltipHandler {
           placement: Placement.bottom,
           rect
         },
-        style: { arrowMark: false }
+        disappearDelay: table.internalProps.tooltip.overflowTextTooltipDisappearDelay ?? 0,
+        style: table.theme.tooltipStyle
       };
     } else if (table.internalProps.tooltip?.isShowOverflowTextTooltip) {
       const overflowText = table.getCellOverflowText(col, row);
@@ -210,6 +211,7 @@ export class TooltipHandler {
             placement: Placement.bottom,
             rect
           },
+          disappearDelay: table.internalProps.tooltip.overflowTextTooltipDisappearDelay ?? 0,
           style: table.theme.tooltipStyle
         };
       }